[UPD] now all menu commands is the same class

+++ src/tasks/task_7/Sort.java
@@ -6,7 +6,7 @@ public final class Sort extends ArrayPI {
         for (int i = 0; i < n - 1; i++) {
             for (int j = 0; j < n - i - 1; j++) {
                 if (array[j] > array[j + 1]) {
-                    int temp = array[j];
+                    double temp = array[j];
                     array[j] = array[j + 1];
                     array[j + 1] = temp;
                 }
@@ -19,7 +19,7 @@ public final class Sort extends ArrayPI {
         for (int i = 0; i < n - 1; i++) {
             for (int j = 0; j < n - i - 1; j++) {
                 if (array[j] < array[j + 1]) {
-                    int temp = array[j];
+                    double temp = array[j];
                     array[j] = array[j + 1];
                     array[j + 1] = temp;
                 }
